backend/requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
asyncpg==0.26.0
bcrypt==3.2.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
distlib==0.3.4
dnspython==2.2.1
docopt==0.6.2
email-validator==1.2.1
fastapi==0.78.0
filelock==3.7.1
h11==0.13.0
identify==2.5.1
idna==3.3
iso8601==1.0.2
nodeenv==1.7.0
passlib==1.7.4
pipreqs==0.4.11
platformdirs==2.5.2
pre-commit==2.19.0
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pypika-tortoise==0.1.6
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tortoise-orm==0.19.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
virtualenv==20.15.1
yarg==0.1.9
